Changed structure for how to paint lidar data
